pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "research-crew"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "crewai>=0.80.0",
    "anthropic>=0.40.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
]

[tool.crewai]
type = "crew"

[tool.hatch.build.targets.wheel]
packages = ["src/research_crew"]
